refactor: update vendor overview layout and components, integrate new title and description, and add comments section
