--- apps/app/src/app/[locale]/(app)/(dashboard)/[orgId]/vendors/[vendorId]/components/charts/risk-overview.tsx
@@ -1,13 +1,0 @@
-import { RisksByDepartment } from "./risks-by-department";
-import { RisksByStatus } from "./risks-by-status";
-
-
-
-export function RiskOverview() {
-  return (
-    <>
-      <RisksByStatus />
-      <RisksByDepartment />
-    </>
-  );
-}
